+++ src/library/DIPS.Mobile.UI/Components/ListItems/Extensions/CheckmarkListItem.cs
@@ -30,8 +30,8 @@ public partial class CheckmarkListItem : ListItem, ISelectable
         IconOptions = m_iconOptions;
         
         //Forces the title to take full width
-        TitleOptions = new TitleOptions() {Width = GridLength.Star}; 
-        InLineContentOptions = new InLineContentOptions() {Width = GridLength.Auto};
+        TitleOptions.Width = GridLength.Star;
+        InLineContentOptions.Width = GridLength.Auto;
         
 #if __ANDROID__
         m_iconOptions.Color = ISelectable.s_tintColor;