+++ src/library/DIPS.Mobile.UI/Converters/ValueConverters/InLineContentTypeConverter.cs
@@ -1,5 +1,7 @@
 using System.ComponentModel;
 using System.Globalization;
+using DIPS.Mobile.UI.Resources.Styles;
+using DIPS.Mobile.UI.Resources.Styles.Label;
 
 namespace DIPS.Mobile.UI.Converters.ValueConverters;
 
@@ -10,7 +12,7 @@ public class InLineContentTypeConverter : TypeConverter
         if (value is not string text)
             return base.ConvertFrom(context, culture, value);
 
-        return new Components.Labels.Label { Text = text };
+        return new Components.Labels.Label { Text = text, Style = Styles.GetLabelStyle(LabelStyle.ValueInlineWithKey) };
     }